+++ core/vexx/v6/absorb.ts
@@ -1,0 +1,8 @@
+import { VexxNodeMatrix } from "../node";
+import { Vexx6NodeType } from "./type";
+
+export class VexxNodeAbsorb extends VexxNodeMatrix {
+  constructor() {
+    super(Vexx6NodeType.ABSORB);
+  }
+}